ref/tool/image_convert/requirements.min.txt
opencv-python
numpy
pytesseract
pypotrace
pillow
